+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Added the ability use local or remote state.
+
 ## v0.1.0 (2014-11-01)
 
 - Initially released.